requirements.txt
Pillow>=9.0.0
cryptography>=40.0.0
flask
qrcode
reedsolo>=1.6.1
pytest>=7.0.0
